Integrated Roast Chicken Prototype and imported some Potato Assets
Packaged and imported my roast chicken scene. The assets for the scene are in the roast chicken prototype resources folder. Also fixed the rendering for my objects to use URP as they used standard shader before. Also replaced my psyche model with the one in the project. Also made a folder for some of the potato object resources as they were in my roast chicken prototype. Made some tweaks to the roast chicken prototype including adding a disclaimer and adjusting lighting and camera position.