Completed Comparison Menu and fixed quiz uncapped next/back buttons in quiz
Added if clause to backandnextbuttons script that checks if the scene is within a certain range to avoid loading non quiz-question scenes.
Switched exit button to a credits button that brings you to credits.
Added functionality to home button in credits to go back to home scene.
Added the other comparisons to the build settings.